re-added files that were removed by the revert
