+++ react/modules/enhancedEcommerceEvents.ts
@@ -1,16 +1,14 @@
 import updateEcommerce from './updateEcommerce'
 import {
-  Order,
   PixelMessage,
   ProductOrder,
-  Impression,
-  CartItem,
   AddToCartData,
   RemoveToCartData,
   Seller,
   ProductViewReferenceId,
 } from '../typings/events'
-import { AnalyticsEcommerceProduct, MaybePrice } from '../typings/gtm'
+import { MaybePrice } from '../typings/gtm'
+import { getCategory, getStoreName, getPurchaseObjectData, getProductObjectData, getProductImpressionObjectData, getCheckoutProductObjectData } from '../utils'
 
 function getSeller(sellers: Seller[]) {
   const defaultSeller = sellers.find(seller => seller.sellerDefault)
@@ -118,8 +116,6 @@ export async function sendEnhancedEcommerceEvents(e: PixelMessage) {
       } catch {
         price = undefined
       }
-
-      const storeName = properties.find((property: { name: string }) => property.name === "Store")
       
       const includePrice: MaybePrice = {}
 
@@ -135,7 +131,7 @@ export async function sendEnhancedEcommerceEvents(e: PixelMessage) {
             products: [
               {
                 brand,
-                affiliation: storeName.values[0],
+                affiliation: getStoreName(properties)?.values[0],
                 category: getCategory(categories),
                 id: productId,
                 variant: sku.itemId,
@@ -195,6 +191,7 @@ export async function sendEnhancedEcommerceEvents(e: PixelMessage) {
           currencyCode: e.data.currency,
           remove: {
             products: items.map(item => ({
+              affiliation: item.affiliation,
               brand: item.brand,
               category: item.category,
               id: item.productId,
@@ -326,101 +323,4 @@ export async function sendEnhancedEcommerceEvents(e: PixelMessage) {
       break
     }
   }
-}
-
-function getPurchaseObjectData(order: Order) {
-  return {
-    affiliation: order.transactionAffiliation,
-    coupon: order.coupon ? order.coupon : null,
-    id: order.orderGroup,
-    revenue: order.transactionTotal,
-    shipping: order.transactionShipping,
-    tax: order.transactionTax,
-  }
-}
-
-function getProductObjectData(product: ProductOrder) {
-  const productName = getProductNameWithoutVariant(
-    product.name,
-    product.skuName
-  )
-
-  return {
-    brand: product.brand,
-    category: product.categoryTree?.join('/'),
-    id: product.id, // Product id
-    variant: product.sku, // SKU id
-    name: productName, // Product name
-    price: product.price,
-    quantity: product.quantity,
-    dimension1: product.productRefId ?? '',
-    dimension2: product.skuRefId ?? '',
-    dimension3: product.skuName, // SKU name (only variant)
-  }
-}
-
-function getCategory(rawCategories: string[]) {
-  if (!rawCategories || !rawCategories.length) {
-    return
-  }
-
-  return removeStartAndEndSlash(rawCategories[0])
-}
-
-// Transform this: "/Apparel & Accessories/Clothing/Tops/"
-// To this: "Apparel & Accessories/Clothing/Tops"
-function removeStartAndEndSlash(category?: string) {
-  return category?.replace(/^\/|\/$/g, '')
-}
-
-function getProductImpressionObjectData(list: string) {
-  return ({ product, position }: Impression) => ({
-    brand: product.brand,
-    category: getCategory(product.categories),
-    id: product.productId, // Product id
-    variant: product.sku.itemId, // SKU id
-    list,
-    name: product.productName,
-    position,
-    price: product.sku.seller.commertialOffer.Price,
-    dimension1: product.productReference ?? '',
-    dimension2: product.sku.referenceId?.Value ?? '',
-    dimension3: product.sku.name, // SKU name (variation only)
-  })
-}
-
-function getCheckoutProductObjectData(
-  item: CartItem
-): AnalyticsEcommerceProduct {
-  const productName = getProductNameWithoutVariant(item.name, item.skuName)
-
-  return {
-    id: item.productId, // Product id
-    variant: item.id, // SKU id
-    name: productName, // Product name without variant
-    category: Object.keys(item.productCategories ?? {}).reduce(
-      (categories, category) =>
-        categories ? `${categories}/${category}` : category,
-      ''
-    ),
-    brand: item.additionalInfo?.brandName ?? '',
-    price: item.sellingPrice / 100,
-    quantity: item.quantity,
-    dimension1: item.productRefId ?? '',
-    dimension2: item.referenceId ?? '', // SKU reference id
-    dimension3: item.skuName,
-  }
-}
-
-function getProductNameWithoutVariant(
-  productNameWithVariant: string,
-  variant: string
-) {
-  const indexOfVariant = productNameWithVariant.lastIndexOf(variant)
-
-  if (indexOfVariant === -1 || indexOfVariant === 0) {
-    return productNameWithVariant
-  }
-
-  return productNameWithVariant.substring(0, indexOfVariant - 1) // Removes the variant and the whitespace
 }
